Move listener to dedicated file
